refactor: add `compileCppTest` function from steps in `src/bin.ts`
