+++ src/test/index.ts
@@ -1,0 +1,1 @@
+export { compileCppTest } from "./cpp.js";
